+++ how-to/workspace-platform-starter/client/types/module/shapes/analytics-shapes.d.ts
@@ -27,6 +27,10 @@ export interface AnalyticsProviderOptions extends ModuleList {
  */
 export declare const PLATFORM_ANALYTICS_SOURCE = "WorkspacePlatform";
 /**
+ * Additional source for module events.
+ */
+export declare const MODULE_ANALYTICS_SOURCE = "Module";
+/**
  * The data for the analytics events that need to be handled. Extends the platform AnalyticsEvent with additional data
  * source.
  */
@@ -34,9 +38,25 @@ export interface PlatformAnalyticsEvent extends Omit<AnalyticsEvent, "source"> {
 	/**
 	 * The source of the event.
 	 */
-	source: typeof PLATFORM_ANALYTICS_SOURCE | AnalyticsSource;
+	source: typeof PLATFORM_ANALYTICS_SOURCE | typeof MODULE_ANALYTICS_SOURCE | AnalyticsSource;
 	/**
 	 * The timestamp for the event.
 	 */
 	timestamp: Date;
+}
+/**
+ * The data for the analytics events that need to be handled. Extends the platform AnalyticsEvent but enforces source as
+ * Module. Modules can use type (to specify module id) and use action, value and data to provide module specific information if required.
+ */
+export type ModuleAnalyticsEvent = Omit<PlatformAnalyticsEvent, "source" | "entityId">;
+/**
+ * Provides a client for handling analytics events.
+ */
+export interface AnalyticsClient {
+	/**
+	 * Handle a list of analytics events.
+	 * @param events The events to handle.
+	 * @returns Nothing.
+	 */
+	handleAnalytics(events: ModuleAnalyticsEvent[]): Promise<void>;
 }